==> notebooks/02_transformed_raw_data_into_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')

rides.head(20)

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0: 'rides'}, inplace=True)

agg_rides

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    output = pd.DataFrame()
    for location_id in tqdm(location_ids):

        # keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]
            
        # quick way to add missing dates with 0 in a Series
        # taken from https://stackoverflow.com/a/19324591
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        
        # add back `location_id` columns
        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])
    
    # move the purchase_day from the index to a dataframe column
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    
    return output

# agg_rides_indexed = agg_rides.set_index(['pickup_hour', 'pickup_location_id'])
# unique_location_ids = rides['pickup_location_id'].unique()

# min_time = agg_rides['pickup_hour'].min()
# max_time = agg_rides['pickup_hour'].max()
# select_times = pd.date_range(min_time, max_time, freq='H')

# agg_multi_index = pd.MultiIndex.from_product([select_times, unique_location_ids], names=['pickup_hour', 'pickup_location_id'])

# agg_final = agg_rides_indexed.reindex(agg_multi_index, fill_value=0).reset_index()
# agg_final

# agg_final[agg_final.rides == 0]

agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots

agg_rides_all_slots[agg_rides_all_slots.rides == 0]

from typing import Optional, List
import plotly.express as px

import warnings
warnings.filterwarnings("ignore", category=FutureWarning)

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None
    ):
    """
    Plot time-series data
    """
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color='pickup_location_id',
        template='none',
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43, 140])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
