==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.to_datetime(datetime.utcnow()).floor('H') # - timedelta(hours=1)
print(f'{current_date=}')
# current_date = pd.Timestamp('2023-02-28 09:00:00')

from src.inference import load_batch_of_features_from_store

features = load_batch_of_features_from_store(current_date)

from src.inference import (
    load_model_from_registry,
    get_model_predictions
)

model = load_model_from_registry()
predictions = get_model_predictions(model, features)

predictions['pickup_hour'] = current_date
predictions['pickup_ts'] = predictions['pickup_hour'].astype(int) // 10**6
predictions

### Save these predictions in the feature store, so they can be later consumed by our Streamlit app

from src.feature_store_api import get_or_create_feature_group
from src.config import FEATURE_GROUP_PREDICTIONS_METADATA, MODEL_NAME

# connect to the feature group
feature_group = get_or_create_feature_group(FEATURE_GROUP_PREDICTIONS_METADATA)

feature_group.insert(predictions, write_options={"wait_for_job": False})
